+++ src/main/java/com/OOL/oolfinance/service/member/MemberService.java
@@ -1,5 +1,6 @@
 package com.OOL.oolfinance.service.member;
 
+import com.OOL.oolfinance.dto.MemberDTO;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -24,13 +25,6 @@ public interface MemberService {
 
     @Transactional
     void updateMemberStatus(Long id);
-    
-    public void save(MemberDTO memberDTO) {
-      //1. dto -> entity 변환
-      //2. repository의 save 메서드 호출
-      Member member = Member.toMember(memberDTO);
-      memberRepository.save(member);
-      //repository의 save 메서드 호출 (조건. entity 객체를 넘겨줘야 함.)
 
-    }
+    void signup(MemberDTO memberDTO);
 }
